+ object prototype will be added to all kernel data types which has  - ptr to destructor of each kernel data type (typically related to system resources)  - destructor willl be invoked when owner thread is destroyed(by error or normal condition)  - it forces to release all occupied resources when they are no longer needed

+++ include/sys/tch_mem.h
@@ -30,7 +30,7 @@ typedef void* tch_memHandle;
  * @brief This is public API for Dynamic Memory Allocation
  */
 struct _tch_mem_ix_t {
-	void* (*alloc)(uint32_t size);
+	void* (*alloc)(size_t size);
 	void (*free)(void*);
 };
